pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cachetools>=6.2.4",
    "litellm>=1.80.13",
    "lxml>=6.0.2",
    "mistletoe>=1.5.1",
    "modal>=1.3.0.post1",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
    "python-fasthtml>=0.12.39",
]

[dependency-groups]
dev = [
    "ipython>=9.9.0",
    "notebook>=7.5.1",
    "plash-cli>=0.3.12",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.11",
]

[tool.pytest.ini_options]
markers = [
    "slow: long-running tests (skipped by default in ./dev test)",
]
filterwarnings = [
    "ignore::UserWarning:pydantic",
]

[tool.ruff]
line-length = 150
exclude = ["junk.ipynb"]
